# calendly-clone-api/docs/en/api.yaml
openapi: 3.0.0
info:
  title: Calendly Clone API
  description: API for scheduling and managing appointments (English documentation)
  version: 1.0.0
servers:
  - url: /
    description: API endpoints
paths:
...
  /schedules:
    post:
      security:
        - BearerAuth: []
      tags:
        - Schedules
      summary: Create a new schedule
      description: Create a new schedule for a user with their availability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleInput'
            example:
              userId: "user-123e4567-e89b-12d3-a456-426614174000"
              availability:
                - day: monday
                  startTime: "09:00"
                  endTime: "17:00"
                - day: tuesday
                  startTime: "09:00"
                  endTime: "17:00"
                - day: wednesday
                  startTime: "09:00"
                  endTime: "17:00"
                - day: thursday
                  startTime: "09:00"
                  endTime: "17:00"
                - day: friday
                  startTime: "09:00"
                  endTime: "16:00"
      responses:
        '201':
          description: Schedule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
              example:
                id: "sched-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                availability:
                  - day: monday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: tuesday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: wednesday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: thursday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: friday
                    startTime: "09:00"
                    endTime: "16:00"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid input"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Internal server error"

    get:
      security:
        - BearerAuth: []
      tags:
        - Schedules
      summary: List all schedules
      description: Retrieve a list of all schedules.
      responses:
        '200':
          description: List of schedules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
              example:
                - id: "sched-123e4567-e89b-12d3-a456-426614174000"
                  userId: "user-123e4567-e89b-12d3-a456-426614174000"
                  availability:
                    - day: monday
                      startTime: "09:00"
                      endTime: "17:00"
                    - day: tuesday
                      startTime: "09:00"
                      endTime: "17:00"
                    - day: wednesday
                      startTime: "09:00"
                      endTime: "17:00"
                - id: "sched-223e4567-e89b-12d3-a456-426614174001"
                  userId: "user-223e4567-e89b-12d3-a456-426614174001"
                  availability:
                    - day: monday
                      startTime: "10:00"
                      endTime: "18:00"
                    - day: wednesday
                      startTime: "10:00"
                      endTime: "18:00"
                    - day: friday
                      startTime: "10:00"
                      endTime: "18:00"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /schedules/{userId}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Schedules
      summary: Get user's schedule
      description: Retrieve the schedule for a specific user by their ID.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user whose schedule to retrieve.
      responses:
        '200':
          description: Schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
              example:
                id: "sched-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                availability:
                  - day: monday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: tuesday
                    startTime: "09:00"
                    endTime: "17:00"
                  - day: wednesday
                    startTime: "09:00"
                    endTime: "17:00"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required to view schedule"
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Schedule not found for user user-123e4567"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Internal server error while retrieving schedule"
    patch:
      security:
        - BearerAuth: []
      tags:
        - Schedules
      summary: Partially update a user's schedule
      description: Update specific fields of a user's schedule by their ID.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user whose schedule to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePatch'
            example:
              availability:
                - day: monday
                  startTime: "10:00"
                  endTime: "18:00"
                - day: wednesday
                  startTime: "10:00"
                  endTime: "18:00"
                - day: friday
                  startTime: "10:00"
                  endTime: "16:00"
      responses:
        '200':
          description: Schedule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
              example:
                id: "sched-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                availability:
                  - day: monday
                    startTime: "10:00"
                    endTime: "18:00"
                  - day: wednesday
                    startTime: "10:00"
                    endTime: "18:00"
                  - day: friday
                    startTime: "10:00"
                    endTime: "16:00"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid time format. Start time must be before end time."
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Schedule not found for user with ID user-123e4567"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: []
      tags:
        - Schedules
      summary: Delete a schedule
      description: Delete the schedule for a specific user by their ID.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user whose schedule to delete.
      responses:
        '204':
          description: Schedule deleted successfully
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Internal server error"
...
components:
  schemas:
...
    Schedule:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        availability:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
      required:
        - id
        - userId
        - availability
    ScheduleInput:
      type: object
      properties:
        userId:
          type: string
        availability:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
      required:
        - userId
        - availability
    TimeSlot:
      type: object
      properties:
        day:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
      required:
        - day
        - startTime
        - endTime
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad Request"
      required:
        - message
...
    SchedulePatch:
      type: object
      properties:
        availability:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
      minProperties: 1  # At least one field must be provided
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token